Create README.md
One of the key factors in determining a country's relative level of economic health is its currency exchange rate. Every free market economy in the world depends on trade, which is significantly influenced by exchange rates. While there are many direct and indirect factors which affect currency exchange rates, the inflation rate of a country and gross domestic product are the most prominent ones. The aim of this project is analyzing the relationship between exchange rates, GDP and inflation rates and checking if we can forecast the economic crisis using effective visualizations.